# pp-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CreateGameRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  handler.EndGameSessionRequest:
    properties:
      score:
        type: integer
    required:
    - score
    type: object
...
  handler.GameResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  handler.GameResponseScore:
    properties:
      id:
        type: string
      player_username:
        type: string
      recorded_at:
        type: string
      score:
        type: integer
      session_id:
        type: string
    type: object
  handler.GameSessionResponse:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      game_id:
        type: string
      id:
        type: string
      player_username:
        type: string
      start_time:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  handler.MaintenanceScheduleResponse:
    properties:
      created_at:
        type: string
      end:
        type: string
      id:
        type: string
      message:
        type: string
      start:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  handler.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
...
  handler.ScheduleRequest:
    properties:
      endTime:
        type: string
      message:
        type: string
      startTime:
        type: string
    required:
    - endTime
    - startTime
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pitturu.com
    name: PittuRu Team
    url: https://pitturu.com
  description: PittuRu 게임 플랫폼 REST API 서버
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PittuRu API
  version: "1.0"
paths:
...
  /admin/maintenance:
    get:
      description: Retrieves the currently scheduled maintenance.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MaintenanceScheduleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Get scheduled maintenance
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Schedules a new maintenance window.
      parameters:
      - description: Maintenance details
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/handler.ScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.MaintenanceScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Schedule maintenance
      tags:
      - Admin
  /admin/maintenance/{id}:
    delete:
      description: Cancels a scheduled maintenance by its ID.
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Cancel maintenance
      tags:
      - Admin
...
  /game-sessions/{session_id}/end:
    put:
      consumes:
      - application/json
      description: Ends an active game session and submits the player's score.
      parameters:
      - description: Game Session ID
        in: path
        name: session_id
        required: true
        type: string
      - description: Score details
        in: body
        name: score
        required: true
        schema:
          $ref: '#/definitions/handler.EndGameSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GameSessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: End a game session
      tags:
      - Game
  /games:
    get:
      description: Retrieves a list of all game definitions.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.GameResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: List all games
      tags:
      - Game
    post:
      consumes:
      - application/json
      description: Creates a new game definition (e.g., OX Quiz, Puzzle Game).
      parameters:
      - description: Game details
        in: body
        name: game
        required: true
        schema:
          $ref: '#/definitions/handler.CreateGameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.GameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Create a new game
      tags:
      - Game
  /games/{game_id}:
    get:
      description: Retrieves a single game definition by its unique ID.
      parameters:
      - description: Game ID
        in: path
        name: game_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GameResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Get game by ID
      tags:
      - Game
  /games/{game_id}/scores:
    get:
      description: Retrieves a list of game scores for a specific game, ordered by
        score.
      parameters:
      - description: Game ID
        in: path
        name: game_id
        required: true
        type: string
      - description: Limit the number of scores returned
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.GameResponseScore'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: List game scores by game ID
      tags:
      - Game
  /games/{game_id}/sessions:
    post:
      consumes:
      - application/json
      description: Creates a new session for a specific game.
      parameters:
      - description: Game ID
        in: path
        name: game_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.GameSessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Create a new game session
      tags:
      - Game
...
  /users/{username}/scores:
    get:
      description: Retrieves a list of game scores for a specific player, ordered
        by recorded time.
      parameters:
      - description: Player username
        in: path
        name: username
        required: true
        type: string
      - description: Limit the number of scores returned
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.GameResponseScore'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: List game scores by player
      tags:
      - Game
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"